#2263 Split LFRicScalarArgs
