+++ src/psyclone/domain/lfric/lfric_invoke.py
@@ -93,7 +93,7 @@ class LFRicInvoke(Invoke):
 
         # Import here to avoid circular dependency
         # pylint: disable=import-outside-toplevel
-        from psyclone.dynamo0p3 import (LFRicScalarArgs, DynStencils,
+        from psyclone.dynamo0p3 import (DynStencils,
                                         DynFunctionSpaces, DynDofmaps,
                                         LFRicFields, DynLMAOperators,
                                         DynCMAOperators, DynBasisFunctions,
@@ -102,6 +102,7 @@ class LFRicInvoke(Invoke):
                                         DynCellIterators, DynReferenceElement,
                                         LFRicMeshProperties, LFRicLoopBounds,
                                         DynGlobalSum)
+        from psyclone.domain.lfric import LFRicScalarArgs
         self.scalar_args = LFRicScalarArgs(self)
 
         # Initialise our Invoke stencil information
