+++ src/psyclone/tests/domain/lfric/lfric_scalar_mdata_test.py
@@ -45,9 +45,8 @@ import os
 import pytest
 import fparser
 from fparser import api as fpapi
-from psyclone.domain.lfric import LFRicArgDescriptor
-from psyclone.dynamo0p3 import (DynKern, DynKernMetadata,
-                                LFRicScalarArgs, LFRicConstants)
+from psyclone.domain.lfric import LFRicArgDescriptor, LFRicScalarArgs
+from psyclone.dynamo0p3 import (DynKern, DynKernMetadata, LFRicConstants)
 from psyclone.errors import InternalError, GenerationError
 from psyclone.f2pygen import ModuleGen
 from psyclone.parse.algorithm import parse
